requirements.txt
requests==2.32.0
beautifulsoup4==4.12.3
pygithub==2.4.0
atproto
